Move mocks and interfaces into unique files`
