+++ main.go
@@ -34,28 +34,6 @@ func (e *env) doCreatePost(c *gin.Context) {
 	createPost(ctx, e.ourDB)
 }
 
-type tacitContext interface {
-	bindJSON(obj interface{}) error
-	readBody([]byte) (int, error)
-	json(int, map[string]interface{})
-}
-
-type realTacitContext struct {
-	ginCtx *gin.Context
-}
-
-func (ctx *realTacitContext) bindJSON(obj interface{}) error {
-	panic("method not implemented")
-}
-
-func (ctx *realTacitContext) readBody([]byte) (int, error) {
-	panic("method not implemented")
-}
-
-func (ctx *realTacitContext) json(int, map[string]interface{}) {
-	panic("method not implemented")
-}
-
 func main() {
 	fmt.Println("Hello, World")
 	// defaultHost := "localhost"
