+++ tacitDB.go
@@ -5,7 +5,7 @@ import "github.com/jinzhu/gorm"
 type tacitDB interface {
 	autoMigrate(values ...interface{})
 	create(value interface{})
-	first(out interface{}, where ...interface{})
+	first(out interface{}, where ...interface{}) 
 	where(query interface{}, args ...interface{}) tacitDB
 }
 
@@ -21,7 +21,7 @@ func (db *realTacitDB) create(value interface{}) {
 	db.gormDB.Create(value)
 }
 
-func (db *realTacitDB) first(out interface{}, where ...interface{}) {
+func (db *realTacitDB) first(out interface{}, where ...interface{}){
 	db.gormDB.First(out, where)
 }
 
@@ -29,3 +29,4 @@ func (db *realTacitDB) where(query interface{}, args ...interface{}) tacitDB {
 	db.gormDB = db.gormDB.Where(query, args)
 	return db
 }
+
